requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15,<4.0
psycopg2-binary>=2.9,<3.0
dj-database-url>=2.2,<3.0
python-dotenv>=1.0,<2.0
gunicorn>=22.0,<23.0
